Indexing Implemented
Added all the functions for the required functions associated with creating an index. This includes:
Allow the user to set a field as an “index”.
Creating an index builds a B+ tree on that field for range searching.
Creating a new index does not discard other existing indexes
Multiple indexes can exist simultaneously
The interface should clearly indicate which fields are currently indexed